Use specific components for person facets row
